requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.7.4
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.7
openai==1.7.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
sniffio==1.3.0
tqdm==4.66.3
typing_extensions==4.9.0
youtube-transcript-api==0.6.3
msgraph-sdk